Add QuantumSavoryGeoMakie
