+++ Project.toml
@@ -8,11 +8,9 @@ Combinatorics = "861a8166-3701-5b0c-9a16-15d98fcdc6aa"
 ConcurrentSim = "6ed1e86c-fcaf-46a9-97e0-2b26a2cdb499"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 DocStringExtensions = "ffbed154-4ef7-542d-bbb7-c09d3a79fcae"
-GeoMakie = "db073c08-6b98-4ee5-b6a4-5efafb3259c6"
 Graphs = "86223c79-3864-5bf0-83f7-82e725a168b6"
 IterTools = "c8e1da08-722c-5040-9ed9-7db0dc04731e"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-NaturalEarth = "436b0209-26ab-4e65-94a9-6526d86fea76"
 NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
@@ -29,8 +27,10 @@ SumTypes = "8e1ec7a9-0e02-4297-b0fe-6433085c89f2"
 
 [weakdeps]
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
+GeoMakie = "db073c08-6b98-4ee5-b6a4-5efafb3259c6"
 
 [extensions]
+QuantumSavoryGeoMakie = "GeoMakie"
 QuantumSavoryMakie = "Makie"
 
 [compat]
