+++ ext/QuantumSavoryMakie/QuantumSavoryMakie.jl
@@ -4,14 +4,12 @@ using QuantumSavory
 using Graphs
 using NetworkLayout
 import ConcurrentSim
-import Makie, GeoMakie
+import Makie
 import Makie: Theme, Figure, Axis,
     @recipe, lift, Observable,
     Point2, Point2f, Rect2f,
     scatter!, poly!, linesegments!,
-    DataInspector, GeometryBasics
-using GeoMakie: GeoAxis, image!, poly!
-using NaturalEarth
+    DataInspector
 import QuantumSavory: registernetplot, registernetplot!, registernetplot_axis, resourceplot_axis, showonplot, showmetadata
 
 ##
@@ -316,17 +314,10 @@ function registernetplot_axis(subfig::Makie.GridPosition, registersobservable; i
     registernetplot_axis(Makie.Axis(subfig), registersobservable; infocli, datainspector, kwargs...)
 end
 
-
-"""Generates a default map with country and state boundaries and returns a GeoAxis. The returned GeoAxis can be used as an input for registernetplot_axis."""
-function generate_map(subfig::Makie.GridPosition)
-    ax = GeoAxis(subfig; limits=((-180, 180), (-90, 90)), dest="+proj=longlat +datum=WGS84")
-    countries = naturalearth("admin_0_countries", 110)
-    states = naturalearth("admin_1_states_provinces_lakes", 110)
-    image!(ax, -180..180, -90..90, GeoMakie.earth() |> rotr90; interpolate=false, inspectable=false)
-    poly!(ax, GeoMakie.land(); color=:lightyellow, strokecolor=:black, strokewidth=1, inspectable=false)
-    poly!(ax, GeoMakie.to_multipoly.(countries.geometry), color=:transparent, strokecolor=:black, strokewidth=1, inspectable=false)
-    poly!(ax, GeoMakie.to_multipoly.(states.geometry); color=:transparent, strokecolor=:grey, strokewidth=0.7, inspectable=false)
-    return ax
+function registernetplot_axis(registersobservable; infocli=true, datainspector=true, kwargs...)
+    fig = Figure()
+    ax = Axis(fig[1, 1])
+    registernetplot_axis(ax, registersobservable; infocli, datainspector, kwargs...)
 end
 
 ##
