+++ test/Project.toml
@@ -17,7 +17,6 @@ JumpProcesses = "ccbc3e58-028d-4f4c-8cd5-9ae44345cda5"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 Makie = "ee78f7c6-11fb-53f2-987a-cfe4a2b5a57a"
-NaturalEarth = "436b0209-26ab-4e65-94a9-6526d86fea76"
 NetworkLayout = "46757867-2c16-5918-afeb-47bfcb05e46a"
 Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Quantikz = "b0d11df0-eea3-4d79-b4a5-421488cbf74b"
